First Implementation of World Border.
Packet is now correctly being send to the client using the WorldBorderUtil. Settings are going to need to be played around with.
Some other small cleanup items.

Need to add an option for the expansion offset of the border. This would allow the border to be expanded outward or inward.